+++ checks/raw/errors.go
@@ -23,4 +23,5 @@ var (
 	errInternalBranchNotFound = errors.New("branch not found")
 	errInvalidArgType         = errors.New("invalid arg type")
 	errInvalidArgLength       = errors.New("invalid arg length")
+	errInvalidGitHubWorkflow  = errors.New("invalid GitHub workflow")
 )
